+++ HISTORY.md
@@ -50,6 +50,8 @@
 - _cattrs_ is now linted with [Ruff](https://beta.ruff.rs/docs/).
 - Remove some unused lines in the unstructuring code.
   ([#416](https://github.com/python-attrs/cattrs/pull/416))
+- Fix handling classes inheriting from non-generic protocols.
+  ([#374](https://github.com/python-attrs/cattrs/issues/374))
 
 ## 23.1.2 (2023-06-02)